Add h- and v- margin/padding shortcuts
This adds four new mixins (h-margin, v-margin, h-padding, and v-padding)
for setting only the values you want without duplicating them. This is
preferable to using the CSS shorthand (which requires you to enter both
the vertical and horizontal padding), and to separate -left and -right
(or -top and -bottom) values, which requires you to duplicate the value.